Class-ify 'aws_vpc'
