+++ src/modules/0.0.17/aws_ecs_fargate/index.ts
@@ -31,12 +31,12 @@ import {
 } from './entity'
 import { Context, Crud2, Mapper2, Module2, } from '../../interfaces'
 import {
-  AwsVpcModule,
   awsCloudwatchModule,
   awsEcrModule,
   awsElbModule,
   awsIamModule,
   awsSecurityGroupModule,
+  awsVpcModule,
 } from '..'
 import * as metadata from './module.json'
 import logger from '../../../services/logger'
@@ -476,8 +476,8 @@ export const AwsEcsFargateModule: Module2 = new Module2({
           // Even though we already have the subnet ids, we look for them to avoid having misconfigured resources
           let subnet: Subnet;
           try {
-            subnet = await AwsVpcModule.mappers.subnet.db.read(ctx, sn) ??
-              await AwsVpcModule.mappers.subnet.cloud.read(ctx, sn);
+            subnet = await awsVpcModule.subnet.db.read(ctx, sn) ??
+              await awsVpcModule.subnet.cloud.read(ctx, sn);
             if (!subnet) return undefined;
           } catch (e: any) {
             if (e.Code === 'InvalidSubnetID.NotFound') return undefined;
