+++ src/modules/0.0.17/aws_elb/index.ts
@@ -34,7 +34,7 @@ import {
   TargetGroupIpAddressTypeEnum,
   TargetTypeEnum,
 } from './entity'
-import { AwsVpcModule, } from '../aws_vpc'
+import { awsVpcModule, } from '../aws_vpc'
 import { Context, Crud2, MapperBase, ModuleBase, } from '../../interfaces'
 import { awsAcmListModule, awsSecurityGroupModule } from '..'
 
@@ -291,8 +291,8 @@ class LoadBalancerMapper extends MapperBase<LoadBalancer> {
     out.securityGroups = securityGroups.filter(sg => !!sg);
     out.ipAddressType = lb.IpAddressType as IpAddressType;
     out.customerOwnedIpv4Pool = lb.CustomerOwnedIpv4Pool;
-    const vpc = await AwsVpcModule.mappers.vpc.db.read(ctx, lb.VpcId) ??
-      await AwsVpcModule.mappers.vpc.cloud.read(ctx, lb.VpcId);
+    const vpc = await awsVpcModule.vpc.db.read(ctx, lb.VpcId) ??
+      await awsVpcModule.vpc.cloud.read(ctx, lb.VpcId);
     out.vpc = vpc;
     out.availabilityZones = lb.AvailabilityZones?.map(az => az.ZoneName ?? '') ?? [];
     out.subnets = lb.AvailabilityZones?.map(az => az.SubnetId ?? '') ?? [];
@@ -602,8 +602,8 @@ class TargetGroupMapper extends MapperBase<TargetGroup> {
     out.healthCheckPath = tg.HealthCheckPath;
     out.protocolVersion = tg.ProtocolVersion as ProtocolVersionEnum;
     try {
-      const vpc = await AwsVpcModule.mappers.vpc.db.read(ctx, tg.VpcId) ??
-        await AwsVpcModule.mappers.vpc.cloud.read(ctx, tg.VpcId);
+      const vpc = await awsVpcModule.vpc.db.read(ctx, tg.VpcId) ??
+        await awsVpcModule.vpc.cloud.read(ctx, tg.VpcId);
       if (tg.VpcId && !vpc) return undefined;
       out.vpc = vpc;
     } catch (e: any) {
