+++ src/modules/0.0.17/aws_security_group/index.ts
@@ -20,7 +20,7 @@ import {
 } from '../../../services/aws_macros'
 import { SecurityGroup, SecurityGroupRule, } from './entity'
 import { Context, Crud2, MapperBase, ModuleBase, } from '../../interfaces'
-import { AwsVpcModule } from '../aws_vpc'
+import { awsVpcModule } from '../aws_vpc'
 import { Vpc } from '../aws_vpc/entity'
 import logger from '../../../services/logger'
 
@@ -40,8 +40,8 @@ class SecurityGroupMapper extends MapperBase<SecurityGroup> {
     out.ownerId = sg.OwnerId;
     out.groupId = sg.GroupId;
     if (sg.VpcId) {
-      out.vpc = await AwsVpcModule.mappers.vpc.db.read(ctx, sg.VpcId) ??
-        await AwsVpcModule.mappers.vpc.cloud.read(ctx, sg.VpcId);
+      out.vpc = await awsVpcModule.vpc.db.read(ctx, sg.VpcId) ??
+        await awsVpcModule.vpc.cloud.read(ctx, sg.VpcId);
       if (!out.vpc) throw new Error(`Waiting for VPC ${sg.VpcId}`);
     };
     return out;
@@ -77,7 +77,7 @@ class SecurityGroupMapper extends MapperBase<SecurityGroup> {
         continue;
       }
       if (!e.vpc) {
-        const vpcs: Vpc[] = await AwsVpcModule.mappers.vpc.cloud.read(ctx);
+        const vpcs: Vpc[] = await awsVpcModule.vpc.cloud.read(ctx);
         if (!vpcs.length) {
           throw new Error('Vpcs need to be loaded first');
         }
@@ -170,11 +170,11 @@ class SecurityGroupMapper extends MapperBase<SecurityGroup> {
         if (out.vpc && out.vpc.vpcId && !out.vpc.id) {
           // There may be a race condition/double write happening here, so check if this thing
           // has been created in the meantime
-          const dbVpc = await AwsVpcModule.mappers.vpc.db.read(ctx, out.vpc.vpcId);
+          const dbVpc = await awsVpcModule.vpc.db.read(ctx, out.vpc.vpcId);
           if (!!dbVpc) {
             out.vpc = dbVpc;
           } else {
-            await AwsVpcModule.mappers.vpc.db.create(out.vpc, ctx);
+            await awsVpcModule.vpc.db.create(out.vpc, ctx);
           }
         }
       }
@@ -192,7 +192,7 @@ class SecurityGroupMapper extends MapperBase<SecurityGroup> {
       const securityGroups = await ctx.orm.find(SecurityGroup, opts);
       for (const sg of securityGroups) {
         if (!sg.vpc) {
-          const vpcs: Vpc[] = await AwsVpcModule.mappers.vpc.db.read(ctx);
+          const vpcs: Vpc[] = await awsVpcModule.vpc.db.read(ctx);
           if (!vpcs.length) {
             throw new Error('Vpcs need to be loaded first');
           }
@@ -213,11 +213,11 @@ class SecurityGroupMapper extends MapperBase<SecurityGroup> {
         if (out.vpc && out.vpc.vpcId && !out.vpc.id) {
           // There may be a race condition/double write happening here, so check if this thing
           // has been created in the meantime
-          const dbVpc = await AwsVpcModule.mappers.vpc.db.read(ctx, out.vpc.vpcId);
+          const dbVpc = await awsVpcModule.vpc.db.read(ctx, out.vpc.vpcId);
           if (!!dbVpc) {
             out.vpc = dbVpc;
           } else {
-            await AwsVpcModule.mappers.vpc.db.create(out.vpc, ctx);
+            await awsVpcModule.vpc.db.create(out.vpc, ctx);
           }
         }
       }
