+++ src/modules/0.0.17/aws_vpc/mappers/index.ts
@@ -1,0 +1,6 @@
+export * from './availability_zone'
+export * from './elastic_ip'
+export * from './endpoint_gateway'
+export * from './nat_gateway'
+export * from './subnet'
+export * from './vpc'
